WIP port to feather esp32s3 tft

+++ components/esp32-s3-boy-hal/hw/driver/audio.c
@@ -120,7 +120,7 @@ void audioThread(void *argument)
   audio_cmd_t *p_cmd;
 
 
-  logPrintf("[__] audioThread %d\n", ch);
+  logPrintf("[__] audioThread %d\n", (int)ch);
 
   while(1)
   {
